REFACTOR test suite to contain one test class per file
